Make AGP dependency compileOnly
This in turn removes the ability to use withPluginClasspath in Gradle TestKit which generally interacts poorly resolving dependencies across the plugin classpath and declared dependencies on the classpath of the project under test. Instead, we replicate the end-user configuration much more accurately by installing to a local maven repository and using good 'ol fashioned buildscript classpath.